+++ packages/uring_api/src/_uring_api/uring_api_capi_impl.h
@@ -35,6 +35,7 @@ int UringApiCapi_RingAutoSubmit(PyObject *ring, int *value);
 int UringApiCapi_RingSetAutoSubmit(PyObject *ring, int value);
 int UringApiCapi_RingWorkerAutoSubmit(PyObject *ring, int *value);
 int UringApiCapi_RingSetWorkerAutoSubmit(PyObject *ring, int value);
+int UringApiCapi_RingStats(PyObject *ring, UringApiRingStats *out);
 int UringApiCapi_CompletionCheck(PyObject *completion);
 PyObject *UringApiCapi_CompletionUserData(PyObject *completion);
 int UringApiCapi_CompletionSetUserData(PyObject *completion, PyObject *value);